% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomBoundary-package.R
\docType{package}
\name{CustomBoundary-package}
\alias{CustomBoundary}
\alias{CustomBoundary-package}
\title{CustomBoundary: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
