% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateFiles.R
\name{generateFiles}
\alias{generateFiles}
\title{generateFiles}
\usage{
generateFiles(
  inputFilePath,
  outputFolder,
  regionName,
  states,
  clustering = TRUE,
  wsStressorData = TRUE,
  ...
)
}
\value{
File paths of generated files
}
\description{
generateFiles
}
