% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBoundaryFile.R
\name{getBoundaryFile}
\alias{getBoundaryFile}
\title{Get Boundary File}
\usage{
getBoundaryFile(inputFilePath, outputFolder, region_name)
}
\arguments{
\item{inputFilePath}{file path to the geospatial file containing the region boundary}

\item{outputFolder}{file path to the folder for saving outputs}

\item{region_name}{name of the region}
}
\value{
writes an RDA file containing the region boundary transformed to EPSG 5070: NAD83/Conus Albers to the output folder
}
\description{
Get Boundary File
}
