% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClusterFile.R
\name{getClusterFile}
\alias{getClusterFile}
\title{Get cluster file}
\usage{
getClusterFile(
  outputFolder,
  regionName,
  states,
  pct_var = 60,
  minCOMIDsCluster = 0.2,
  user_numclust = NULL
)
}
\arguments{
\item{outputFolder}{file path containing output files}

\item{regionName}{region name}

\item{states}{vector of states in the analysis region}
}
\value{
writes a cluster files and graphics for a custom region to the outputFolder.

writes a folder with supplemental files like the variable stats file
}
\description{
Get cluster file
}
