% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReachesFile.R
\name{getReachesFile}
\alias{getReachesFile}
\title{Get Reaches File}
\usage{
getReachesFile(outputFolder, region_name, states)
}
\arguments{
\item{outputFolder}{file path to the folder for saving outputs}

\item{region_name}{name of the region}
}
\value{
writes an rda file containing the comids and geometry of flowines within the custom region to the outputFolder
}
\description{
Get Reaches File
}
