% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWSStressFiles.R
\name{getWSStressFiles}
\alias{getWSStressFiles}
\title{Get watershed stressor data and metadata}
\usage{
getWSStressFiles(outputFolder, regionName, states)
}
\arguments{
\item{outputFolder}{file path containing output files}

\item{regionName}{region name}

\item{states}{vector of states in the analysis region}
}
\value{
Writes a csv with watershed stressor data within the analysis region formatted for the CASTool and watershed stressor metadata
}
\description{
Get watershed stressor data and metadata
}
